refactor(backend): extract config, env validation, error handler from index.ts (#43)
- Move Env type + getEnv() to config/env.ts
- Move authOptions to config/auth.ts
- Extract error handler to middleware/error-handler.ts
- Delete utils/iso.ts re-export (#40)
- Split referidos.ts into routes + handlers (#45)
- Add toPublicPremio() transformer (#38)

Closes #40, Closes #43, Closes #45, Closes #38
