+++ backend/src/domain/types/premios.ts
@@ -1,0 +1,7 @@
+export type PublicPremio = {
+  id: string;
+  nombre: string;
+  descripcion: string;
+  puntos_requeridos: number;
+  activo: boolean;
+};
