+++ backend/src/index.ts
@@ -1,5 +1,8 @@
 import { Elysia } from 'elysia';
 
+import { getEnv } from './config/env';
+import { getAuthOptions } from './config/auth';
+import { createErrorHandler } from './middleware/error-handler';
 import { authContextModule } from './modules/auth-context';
 import { registerAuthRoutes } from './modules/auth';
 import { registerServiciosRoutes } from './modules/servicios';
@@ -8,57 +11,11 @@ import { registerProfilesRoutes } from './modules/profiles';
 import { registerReferidosRoutes } from './modules/referidos';
 import { registerPremiosRoutes } from './modules/premios';
 
-type Env = {
-  PORT: number;
-  NODE_ENV: 'development' | 'test' | 'production';
-  JWT_SECRET: string;
-};
-
-function getEnv(): Env {
-  const portRaw = process.env.PORT;
-  const port = portRaw ? Number(portRaw) : 3001;
-
-  if (!Number.isFinite(port) || port <= 0) {
-    throw new Error(`Invalid PORT: ${String(portRaw)}`);
-  }
-
-  const nodeEnvRaw = process.env.NODE_ENV;
-  const nodeEnv = (nodeEnvRaw || 'development') as Env['NODE_ENV'];
-  if (!['development', 'test', 'production'].includes(nodeEnv)) {
-    throw new Error(`Invalid NODE_ENV: ${String(nodeEnvRaw)}`);
-  }
-
-  const jwtSecret = process.env.JWT_SECRET;
-  if (!jwtSecret) {
-    throw new Error('JWT_SECRET is required');
-  }
-
-  return { PORT: port, NODE_ENV: nodeEnv, JWT_SECRET: jwtSecret };
-}
-
 const env = getEnv();
-
-const authOptions = {
-  jwtSecret: env.JWT_SECRET,
-  nodeEnv: env.NODE_ENV,
-} as const;
+const authOptions = getAuthOptions(env);
 
 const app = new Elysia()
-  .onError(({ code, error, set }) => {
-    if (code === 'VALIDATION') {
-      set.status = 422;
-      return { error: 'validation_error' };
-    }
-
-    if (env.NODE_ENV !== 'test') {
-      console.error(`[api] error code=${code}`, error);
-    }
-
-    set.status = 500;
-    return {
-      error: 'internal_server_error'
-    };
-  })
+  .onError(createErrorHandler(env.NODE_ENV))
   .use(authContextModule(authOptions))
   .get('/api/health', () => ({ ok: true }))
   .get('/api/_admin', ({ auth, status }) => {
