+++ backend/src/middleware/error-handler.ts
@@ -1,0 +1,19 @@
+import type { ErrorHandler } from 'elysia';
+
+export function createErrorHandler(nodeEnv: string): ErrorHandler {
+  return ({ code, error, set }) => {
+    if (code === 'VALIDATION') {
+      set.status = 422;
+      return { error: 'validation_error' };
+    }
+
+    if (nodeEnv !== 'test') {
+      console.error(`[api] error code=${code}`, error);
+    }
+
+    set.status = 500;
+    return {
+      error: 'internal_server_error',
+    };
+  };
+}
