--- backend/src/utils/iso.ts
@@ -1,1 +1,0 @@
-export { asIsoString } from '../domain/transformers/iso';
